Move getObject(ResultSet...) from test into server module
